requirements.txt
amply==0.1.4
certifi==2020.12.5
docutils==0.16
joblib==1.0.1
numpy==1.20.1
Pillow==8.1.0
PuLP==2.1
pyparsing==2.4.7
scikit-learn==0.24.1
scipy==1.6.1
sklearn==0.0
threadpoolctl==2.1.0
torch==1.7.1
torchvision==0.8.2
typing-extensions==3.7.4.3
